handle complex authentication process

+++ src/Controller/PostController.php
@@ -26,6 +26,7 @@ class PostController extends AbstractController
     #[Route('/post/new')]
     public function create(Request $request, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $post = new Post();
         $form = $this->createForm(PostType::class, $post);
         $form ->handleRequest($request);
@@ -44,6 +45,7 @@ class PostController extends AbstractController
     #[Route('/post/edit/{id<\d+>}', name:"edit-post")]
     public function update(Request $request, Post $post, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $form = $this->createForm(PostType::class, $post);
         $form ->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -52,14 +54,16 @@ class PostController extends AbstractController
             return $this->redirectToRoute("home");
             //dump($post);
         }
-        return $this->render('form.html.twig', [
-            "post_form" => $form->createView()  
+        return $this->render('post/form.html.twig', [
+            #"post_form" => $form->createView()
+            "form" => $form->createView()  
         ]);
     }
 
     #[Route('/post/delete/{id<\d+>}', name:"delete-post")]
     public function delete(Post $post, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $em = $doctrine->getManager();
         $em->remove($post);
         $em->flush();
@@ -69,6 +73,7 @@ class PostController extends AbstractController
     #[Route('/post/copy/{id<\d+>}', name:"copy-post")]
     public function duplicate(Post $post, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $copyPost = clone $post;
         $em = $doctrine->getManager();
         $em->persist($copyPost);
